Crate problem102

Problem 102 (Triangle containment)

Solution detail

To check if a triangle ABC contains a point P, we need to know two facts:

P is within the triangle iff the signed areas of triangles ABP, BCP and CAP all have the same non-zero sign.

The signed area of triangle XYZ with coordinates (x₁, y₁), (x₂, y₂), (x₃, y₃) is equal to the determinant of the 3x3 matrix having (xᵢ, yᵢ, 1) as rows.

Thus, to check that a triangle ABC contains the origin, we simply need to calculate the signum of the signed areas of triangles ABO, BCO, CAO and check that they are all equal.

Functions

answer
Solve the problem, returning the answer as a String

contains_origin
Check if the given triangle contains the origin.

main

solve
Find the count of the given triangles which contain (strictly) the origin.

Type Definitions

Point

Triangle